Enum lam_emu::bytecode::Instruction

#[repr(C)]pub enum Instruction {
    Halt,
    Move(Value, Register),
    Swap(Register, Register),
    Clear(Register),
    Allocate {
        words: u8,
        keep_registers: u8,
    },
    Deallocate {
        words: u8,
    },
    Label(Label),
    Jump(Label),
    Return,
    Test(Label, Test),
    Badmatch,
    Call(FnCall, FnKind),
    TailCall(FnCall, FnKind),
    MakeLambda {
        first_label: Label,
        module: Atom,
        arity: Arity,
        environment_size: u32,
    },
    ConsList {
        head: Value,
        tail: Value,
        target: Register,
    },
    SplitList {
        list: Register,
        head: Register,
        tail: Register,
    },
    SplitListTail {
        list: Register,
        tail: Register,
    },
    SplitListHead {
        list: Register,
        head: Register,
    },
    GetTupleElement {
        tuple: Register,
        element: u32,
        target: Register,
    },
    Spawn(Spawn),
    Sleep,
    Kill,
    Monitor,
    Send {
        message: Value,
        process: Value,
    },
    PeekMessage {
        on_mailbox_empty: Label,
        message: Register,
    },
    RemoveMessage,
    PidSelf(Register),
}

Variants

Halt

Kill-switch.

A Halt instruction will stop an emulator.

Move(Value, Register)

Working with the Register Machine

Move value or register value to a register

Swap(Register, Register)

Swap the values of two registers

Clear(Register)

Zero a register

Allocate

Working with the Heap

Allocate

Fields of Allocate

words: u8

Amount of words to allocate on the heap

keep_registers: u8

how many registers to preserve NOTE(@ostera): this is currently an artifact of how BEAM byteops * work. This should be split into 2 operations: allocate + clear_many

Deallocate

Deallocate

Fields of Deallocate

words: u8

Label(Label)

Flow-control operations

Define a new label.

Jump(Label)

Jump to a label.

Return

Returns control to the last Continuation Pointer.

Test(Label, Test)

Perform a test and jump to label if it fails.

Badmatch

Could not match a value to a pattern

Call(FnCall, FnKind)

Calling functions

Perform a function call with or without allocating a new stack frame.

TailCall(FnCall, FnKind)

MakeLambda

Working with Values

Create a boxed lambda and place it on Global(0). When called with Call(Lambda { register }) execution will start at first_label.

Fields of MakeLambda

first_label: Labelmodule: Atomarity: Arityenvironment_size: u32

ConsList

Cons head onto tail and place it in the target register

Fields of ConsList

head: Valuetail: Valuetarget: Register

SplitList

Deconstruct a list from list and place its head in register head and its tail in register tail

Fields of SplitList

list: Registerhead: Registertail: Register

SplitListTail

Split a list and put the tail in a register

Fields of SplitListTail

list: Registertail: Register

SplitListHead

Split a list and put the head in a register

Fields of SplitListHead

list: Registerhead: Register

GetTupleElement

Copy a tuple element onto a specific registry

Fields of GetTupleElement

tuple: Registerelement: u32target: Register

Spawn(Spawn)

Processes

Creates a new process looking for a lambda at Register, and puts the Pid on the X(0) register

Sleep

Put a process to sleep. Will be woken up by a new message.

Kill

Monitor

Send

Puts the message into the mailbox of the pid

Fields of Send

message: Valueprocess: Value

PeekMessage

Check if the mailbox is empty and jump to a label if it is

Fields of PeekMessage

on_mailbox_empty: Labelmessage: Register

RemoveMessage

Removes the currently selected message in the mailbox

PidSelf(Register)

Puts the identifier of the current process in a register